Partially Complete Dependency Lookup
CDFK will now check if the arguments
passed to a task is an ExecFutureWrapper.
If none of the arguments are ExecFutureWrappers then
the task can be executed as normal, since it has no depedencies
however if the task has ExecFutureWrapper(s) in its
arguments then it has a depedency and is not launched.
Still need to implement the case where an ExecFutureWrapper is
passed as an argument but happens to be already complete, in this
case the task can be launched. Will Likely Modify AppFuture and replace
ExecFutureWrapper with AppFuture since they serve the same purpose.
Will have to modify AppFutre in this case.