Changed semantics of task->depends
t->depends now contains the task ids of tasks dependent on
task t. t->depsize is the size of the array t->depends.
t->depcount is the number of depdencies in t->depends(a measure
of occupancy) and t->depon is the number of tasks that t is depedent on.
If this number is 0 then it can be launched. adddep_task was modified to
reflect these changes.