Changed dfk from array to hashtable
Works perfectly fine when the number of buckets is equal to the number
of tasks but, occasionally segfaults when there are more task then buckets
in the hashtable. The segfault happens in the python function do_richcompare.
Function: https://github.com/python/cpython/blob/main/Objects/object.c#L776.
The functions takes 2 PyObject* as arguments, so trying to see if there
is a way to figure what objects are being passed. Could be a refernce
counting issue when allocating new layers of the hashtable.
Also could be an issue with memset when allocting new layers of the hashtable.