Changed delete and added it too insert_tasktable
Its commented out as of now so impact can be measured, but
the insert_tasktable will look at the tasktable to see if a task
is done. If a task is done then insert_tasktable will try to
delete that task and return the position of the deleted task as a
place to put a new task. Also changed how delete functions, now
delete_task does not lookup the task, it must be provided an index
and a depth into the tasktable